feat(viewer): move render-contract types into brepjs-viewer
Lift MeshData/FaceGroup/EdgeGroup/FaceInfo/EdgeInfo/ViewMode/ScreenPos plus
VIEW_NAMES/ViewName into packages/brepjs-viewer/src/types.ts and re-export from
src/index.ts. Add brepjs-viewer as an apps/playground dependency with a tsconfig
paths mapping so the type re-exports resolve from source before dist exists, and
re-export the contract types from the playground stores/worker protocol.

Pin @types/react to 19.2.14 at the root so @react-three/fiber's JSX
augmentation resolves the same react/jsx-runtime the playground uses, and
resolve the playground vite react alias dynamically since npm now hoists react.

Gate: npm run build --workspace=apps/playground (tsc -b + vite build) passes.
