+++ apps/playground/package.json
@@ -21,6 +21,7 @@
     "@vercel/analytics": "2.0.1",
     "brepjs": "*",
     "brepjs-opencascade": "*",
+    "brepjs-viewer": "*",
     "lz-string": "1.5.0",
     "monaco-editor": "0.55.1",
     "react": "19.2.6",
