+++ package.json
@@ -250,6 +250,7 @@
     "@eslint/js": "10.0.1",
     "@size-limit/file": "12.1.0",
     "@types/opentype.js": "1.3.9",
+    "@types/react": "19.2.14",
     "@vitest/coverage-v8": "4.1.6",
     "brepjs-opencascade": "*",
     "brepkit-wasm": "2.101.2",
